Stop loading net_storage.rake twice
