+++ lib/capistrano/net_storage.rb
@@ -1,2 +1,31 @@
-require 'capistrano/net_storage/all'
-load File.expand_path('../tasks/net_storage.rake', __FILE__)
+module Capistrano
+  module NetStorage
+    class << self
+      attr_reader :config
+
+      def setup!(config:)
+        @config = config
+      end
+
+      def transport
+        config.transport_class.new
+      end
+
+      def archiver
+        config.archiver_class.new
+      end
+
+      def scm
+        config.scm_class.new
+      end
+
+      def cleaner
+        config.cleaner_class.new
+      end
+
+      def bundler
+        config.bundler_class.new
+      end
+    end
+  end
+end
