+++ lib/capistrano/net_storage/error.rb
@@ -1,6 +1,6 @@
 module Capistrano
   module NetStorage
-    class Error < StandardError
+    class Error < StandardError # just for backward compatibility
     end
   end
 end
